Add files via upload
added raw - in work - version of the Check ESP tool. It contains a setup for auto closing the dialogs. This is for placing it into the startup items to regulary check ESPs for a Microsoft bootloader file containing certificates. Also it scans the nvram for certificates when a Mac Pro is detected. This can be configured, if the csr option "allow untrusted kexts" is not wanted.